refactor(entrypoints): 辅进程入口统一后缀
