+++ bot_embed_aux.py
