+++ bot_work_aux.py
