+++ docs/maintainer/deploy/single-process.md
@@ -1,6 +1,6 @@
 # 单进程部署
 
-单进程承载消息主路径，并自动维护 **work aux**；本机 Embedding 也可另起 `bot_embed`。适合本地开发、自用与默认生产路径；多数站点先用本形态，仅在验证过瓶颈或需要强隔离时再上 [分片](sharded.md)。
+单进程承载消息主路径，并自动维护 **work aux**；本机 Embedding 也可另起 `bot_embed_aux`。适合本地开发、自用与默认生产路径；多数站点先用本形态，仅在验证过瓶颈或需要强隔离时再上 [分片](sharded.md)。
 
 ## 特点
